feat: implement reusing already stored queues under more general queries by NfzQueuesCacheService#get()
Now get() will use stored queues not only when called with exatly the
same query that was stored along with those queues, but also when called
with a query that is different, but more specific than the original
query.

It works because we always fetch all data/pages from NFZ API for a
particular request. E.g. when somebody requests:
  benefit = Dermatologist
  province = Silesia
we store under that query in database ALL queues to dermatologists in
Silesia.

Therefore, when we later call get() with a query that e.g. additionally
specifies locality = KATOWICE (which is a city in Silesia), there is no
need to fetch data again from NFZ API servers, as we can simply filter
already stored queues and return only those that additionally have
locality = KATOWICE.